why am i writing commits to myself
yeah logi nodes done
some minor adjustments
fixed vic attributes
