+++ a3ae_optre/addons/templates/CfgLogistics.hpp
@@ -1,0 +1,18 @@
+//redefine the addon macro so we can use a3a tools for nodes and cargo
+#ifdef ADDON
+#undef ADDON
+#endif
+#define ADDON A3A_Logistics
+
+class DOUBLES(ADDON,Nodes) {
+    class TRIPLES(ADDON,Nodes,Base); //import base class from antistasi
+    #include "Logistics\OptreNodes.hpp" //include the nodes from the different mods here 
+};
+class DOUBLES(ADDON,Cargo) {
+    class TRIPLES(ADDON,Cargo,Base); //import base class from antistasi
+    #include "Logistics\OptreCargo.hpp" //include the cargo from the different mods here
+};
+
+//redefine the proper addon macro
+#undef ADDON
+#define ADDON DOUBLES(PREFIX,COMPONENT)
