--- a3ae_optre/addons/templates/Logistics/Nodes/OptreNodes.hpp
@@ -1,64 +1,0 @@
-/*
-past output from the function: A3A_Logistics_fnc_generateHardPoints
-*/
-
-/*
-    Function: A3A_Logistics_fnc_generateHardPoints
-    Author: [Håkon]
-    [Description]
-        Generates a rough node array for you based on the visual start and end of the intended cargo plane,
-        in addition visual guides are rendered on screen to give you an idea of where it is relative to you.
-        visuals last for 60 seconds.
-
-        These are:
-            Green dot   : start position of cargo plane (only visible if you can see the position)
-            Red dot     : end position of cargo plane (only visible if you can see the position)
-            white line  : cargo plane bounds
-            white dots  : Node return positions
-
-    Arguments:
-    0. <Object> Vehicle your generating the nodes for
-    1. <Array>  Model relativ position of cargo plane start position
-    2. <Int>    The lenght of the cargo plane
-    3. <Bool>   Return preped for model based definition instead of class based
-
-    Return Value:
-    <text> generated node class to be pasted in to the config
-
-    Scope: Clients
-    Environment: unscheduled
-    Public: [Yes]
-    Dependencies:
-
-    Example: [cursorTarget, [0,-0.7,-0.7], 2.1] call A3A_Logistics_fnc_generateHardPoints;
-*/
-
-/*
-example output
-*/
-class Example_A3_Soft_F_Quadbike_01_Quadbike_01_F_p3d : TRIPLES(ADDON,Nodes,Base)
-{
-    class Nodes
-    {
-        class Node1
-        {
-            offset[] = {0,-0.9,-0.453};
-            seats[] = {0};
-        };
-    };
-};
-
-class OPTRE_DME_M12_FAV : TRIPLES(ADDON,Nodes,Base)
-{
-        class Nodes
-    {
-        class Node1
-        {
-            offset[] = {0,-1.1,-0.7};
-        };
-        class Node2
-        {
-            offset[] = {0,-1.9,-0.7};
-        };
-    };
-};
