--- a3ae_optre/addons/templates/Logistics/config.cpp
@@ -1,4 +1,0 @@
-
-class A3A {
-    #include "CfgLogistics.hpp"
-};
