+++ a3ae_optre/addons/templates/config.cpp
@@ -14,10 +14,7 @@ class CfgPatches {
     };
 };
 
-
 class A3A {
     #include "Templates.hpp"
     #include "CfgLogistics.hpp" //this should be at the end of the config.cpp as it will redefine the macro ADDON
 };
-
-
